Add raw throughput to per internal output file in neper
Currently, if we specify -A, a csv file is generated with per interval stats
(default interval is 1s). The csv file already has a field to record per
interval bytes or transactions.

This change adds a new field to record the raw throughput value per interval into
the csv file. For stream tests, the unit is bytes/s. And for rr tests, the unit
is transactions/s. I believe this is useful to check if there is any
fluctuations in the throughput value during the test.