neper: fix a few compiler warnings
- control_plane.c:166:13: warning: ‘sfd’ may be used uninitialized in this function [-Wmaybe-uninitialized]

This is a false positive. getaddrinfo_or_die crashes the binary if
not at least one item is found that results in sfd = socket(..).

- loop.c:23:18: warning: ‘TEN_MILLISECONDS’ defined but not used [-Wunused-const-variable=]

Never actually used, as far as a `git log -p` shows.

- stats.c:151:33: warning: ‘raw_thruput’ may be used uninitialized in this function [-Wmaybe-uninitialized]

True case. This variable is used to log a diff between iterations of a
loop. Initialized in all but the first iteration, but logging happens
also in this first iteration.

- thread.c:289:32: warning: ‘end’ may be used uninitialized in this function [-Wmaybe-uninitialized]

Slightly confusing, but this is initialized at the same time start gets
its first real value, in `if (start < 0) start = end = i`. Give it a
default initializer.