Allow neper to run in an unprivileged network namespace
In unprivileged network namespaces - that is network namespaces owned by
a non-init user namespace - the sysctls in /proc/sys/net/core are not
accessible.  This caused neper tools to fail on startup since the
procfile_int() call in check_options_common() trips a fatal error if it
can't open /proc/sys/net/core/somaxconn.

Allow neper to run in such a namespace by only making the check if
/proc/sys/net/core/somaxconn is openable.  Since this was the only caller
of procfile_int(), remove it and just open-code the new version within
check_options_common().

Signed-off-by: David Gibson <dgibson@redhat.com>