neper: use one socket per flow in bidirectional stream mode
In bidirectional mode, acks are piggybacked behind data and this
creates unwanted dependencies between forward and reverse flows.

To solve the problem, IN BIDIRECTIONAL STREAM MODE ONLY we use
one tcp socket per direction (the user-specified number of flows
is doubled after option parsing), used as follows:
- client and server always read from all sockets
- client sends only on half of the sockets (those witheven f_id).
  This is done by disabling EPOLLOUT on alternate sockets.
- server starts sending on all sockets, but will stop sending and
  disable EPOLLOUT on sockets on which data is received.
  This is done in stream_handler()

The above allows to have half of the sockets in tx, and half in rx,
without control plane modifications.

For backward compatibility, this is controlled by the --split-bidir
command line option which implies -rw on both sides.

Tested: manual test with --split-bidir and different '-m' values on client and server