histo: replace threshold table with faster bit-based logarithms
histograms store samples in buckets with pseudo logarithmic size.
The previous implementation used a table of thresholds, and binary
search to locate the correct bucket.

This patch replaces the thresholds with the fast pseudo-logarithm
algorithm used in lr-cstats and bpftrace so we can locate the bucket in
a handful of instructions.

This gives memory savings, reduced cache trashing, and better
performance. Tests show that with a hot cache a lookup now takes
less than 2us compared to 20-25us with the previous approach.
Also, we can remove the now-useless neper_histo_factory.

The actual resolution of the buckets is approximately the same as in
the previous implementation (about 1.5%).

On passing, correct a few bugs in the previous implementation:
- resolution was supposed to be 0.25% but due to an implementation bug
  it was around 1% or even bigger at low values, and cause the
  thresholds to become negative
- conversion from double to int for the sample could have unchecked
  overflows.

Tested with tcp_rr and verifying that the distribution and csv files
contain correct values.