histograms: allow arbitrary percentiles
Allow arbitrary percentiles to be specificed, instead of just integer
and p99.9 and p99.99

This also makes the code faster because we can just compute the values
requested instead of all 103 entries.

Any floating point number between 0 and 100 is now accepted, with 999
and 9999 mapped to 99.9 and 99.99 for backward compatibility.

Tested as usual with
 ./tcp_rr -c -H 127.0.0.1 -p1,2,10,50,90,999,9999,100 -A/tmp/x.csv
 and verifying the correct content of the csv file.