histo: only scan necessary buckets when computing percentiles
Computing percentiles is expensive, as it requires scanning all the 4k-8k
buckets used to store samples, and is done for each flow.  Benchmarks show
the original code took an average of 20us per flow, with frequent peaks
in the 60-80us range.

This patch eliminates the cost by not storing samples in buckets if no
percentiles are requested, and otherwise achieves a ~5x reduction by
tracking the range of buckets that contain values in each epoch.

Also change the precision to 6 bits, which halves the cost without much
impact on the results. This value may become a command line flag.

Tested, as usual, by running tcp_rr and verifying the logs and csv