Move TCP_TX_DELAY setsockopt after 3WHS
Setting TCP_TX_DELAY right after socket creation makes SYN/SYN-ACK
packets also
get delayed. This introduces extra amount of time to establish the
connections
when -F is high. This is not ideal and does not reflect the reality of
pacing.

This change moves the TCP_TX_DELAY setting to after 3WHS is finished and
the
socket is in established state.

Author: Wei Wang