neper: rx zerocopy stats
Tested: ran locally with rx zerocopy. Saw that, with the remote using tx
zerocopy, we regularly hit 99% of bytes being read via zerocopy. When
the remote doesn't use tx zerocopy, in which case it tends not to send
the page-sized/aligned data required for rx zerocopy, the percentage is
much lower.