Remove redundant callback logic for --logtostderr enablement.
During the logging_init function call we are already setting log_file to stderr
when --logtostderr file is enabled. So, now there is no need of callback function
that need to be registered like what logtostderr is currently doing.

Tested: Ran make and tests for all type of workload with --logtostderr flag and
error is getting logged to the console(stderr) correctly.