remove unreachable -M flag in tcp_stream
tcp_stream currently defines 2 -M flags. One, which sets SO_MARK, is
unusable because the other definition of -M takes precedence (by virtue
of being earlier in the getopt optstring). So removing it breaks no
users.

Another considered option was to use a glibc extension of getopt that
allows for optional arguments. However, this (a) made for a bunch of
messy code, and (b) has the quirk that optional arguments must part of
the same argv element. So "-M 12" wouldn't work anymore, just "-M12".